requirements.txt
django==3.1.7
mysqlclient==2.0.3
django-oauth-toolkit==1.4.1
djangorestframework==3.12.2
django-cors-middleware==1.5.0
geocoder==1.38.1


